Extract `BlueprintWarningNotice` and `BlueprintIssuesModal` to separate file
